+++ CHANGELOG.md
@@ -17,6 +17,11 @@
     Click to see more.
   </summary>
 
+## :bug: Bug Fix
+
+- `ricos-plugins`
+  - [#4592](https://github.com/wix/ricos/pull/4592) external insert plugins toolbar support toolbar settings
+
 </details>
 
 ## 8.72.64 (Sep 13, 2022)